*: rename files in snake case
closes #27.
